Add weight

+++ learn/trainfile/records/read.py
@@ -2,7 +2,7 @@ def read(file):
     for line in file:
         line = line.strip()
         if line:
-            yield list(map(int, line.split(" ")))
+            yield list(map(float, line.split(" ")))
         else:
             yield None
     yield None
